Extract traverser and emitter from runner
